apps/worker/requirements.txt
python-dotenv==1.0.1
supabase==2.3.4
httpx==0.26.0
google-generativeai==0.4.0
schedule==1.2.1
pydantic==2.6.1
pydantic-settings==2.1.0
